requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cheroot==8.6.0
CherryPy==18.6.1
colorama==0.4.4
cryptography==36.0.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.10
django-appconf==1.0.5
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-libsass==0.9
django-referrer-policy==1.0
django-sass-processor==1.1
djangorestframework==3.13.1
dnspython==2.2.0
docopt==0.6.2
email-validator==1.1.3
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipython==7.16.2
ipython-genutils==0.2.0
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.context==4.1.1
jaraco.functools==3.4.0
jaraco.text==3.7.0
jedi==0.17.2
libsass==0.21.0
Markdown==3.3.6
more-itertools==8.12.0
mysqlclient==2.0.3
oauthlib==3.1.1
parso==0.7.1
phonenumbers==8.12.41
pickleshare==0.7.5
pipreqs==0.4.11
portend==3.0.0
prompt-toolkit==3.0.24
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyQiwiP2P==1.0.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
pywin32==303
rcssmin==1.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
tempora==4.1.2
traitlets==4.3.3
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
yarg==0.1.9
zc.lockfile==2.0
zipp==3.6.0
mutagen~=1.45.1
transliterate~=1.10.2